api/requirements.txt
# Flask Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1

# Database
SQLAlchemy==2.0.23

# Google Places API
googlemaps==4.10.0

# Environment Variables
python-dotenv==1.0.0

# HTTP Requests
requests==2.31.0

# JSON Processing
marshmallow==3.20.1

# Date/Time Handling
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-flask==1.3.0

# Code Quality
flake8==6.1.0
black==23.12.0
